Module Program

+++ 2.cpp
@@ -1,10 +1,10 @@
 #include<stdio.h>
-void mod(int a,int b){
-	if(a%b==0) printf("Perfect Module");
+void mod(float a,float b){
+	if((int)a%(int)b==0) printf("Perfect Module");
 	else printf("Not a perfect module");
 }
 int main(){
-	int a,b;
-	scanf("%d %d",&a,&b);
+	float a,b;
+	scanf("%f %f",&a,&b);
 	mod(a,b);
 }
